refactor and add more examples
