+++ examples/shared/setup.py
