+++ pyproject.toml
@@ -39,29 +39,6 @@ dependencies = [
     "tomli>=2.2.1",
 ]
 
-[project.urls]
-Homepage = "https://github.com/tadata-org/fastapi_mcp"
-Documentation = "https://github.com/tadata-org/fastapi_mcp#readme"
-"Bug Tracker" = "https://github.com/tadata-org/fastapi_mcp/issues"
-"PyPI" = "https://pypi.org/project/fastapi-mcp/"
-"Source Code" = "https://github.com/tadata-org/fastapi_mcp"
-"Changelog" = "https://github.com/tadata-org/fastapi_mcp/blob/main/CHANGELOG.md"
-
-[project.scripts]
-fastapi-mcp = "fastapi_mcp.cli:app"
-
-[tool.hatch.build.targets.wheel]
-packages = ["fastapi_mcp"]
-
-[tool.ruff]
-line-length = 120
-target-version = "py310"
-
-[tool.pytest.ini_options]
-asyncio_mode = "auto"
-testpaths = ["tests"]
-python_files = "test_*.py"
-
 [dependency-groups]
 dev = [
     "mypy>=1.15.0",
@@ -72,3 +49,23 @@ dev = [
     "pytest-cov>=4.1.0",
     "pre-commit>=4.2.0",
 ]
+
+[project.urls]
+Homepage = "https://github.com/tadata-org/fastapi_mcp"
+Documentation = "https://github.com/tadata-org/fastapi_mcp#readme"
+"Bug Tracker" = "https://github.com/tadata-org/fastapi_mcp/issues"
+"PyPI" = "https://pypi.org/project/fastapi-mcp/"
+"Source Code" = "https://github.com/tadata-org/fastapi_mcp"
+"Changelog" = "https://github.com/tadata-org/fastapi_mcp/blob/main/CHANGELOG.md"
+
+[tool.hatch.build.targets.wheel]
+packages = ["fastapi_mcp"]
+
+[tool.ruff]
+line-length = 120
+target-version = "py312"
+
+[tool.pytest.ini_options]
+asyncio_mode = "auto"
+testpaths = ["tests"]
+python_files = "test_*.py"
